Initial Commit
First Commit of the code. It works, but is not pretty at all. Need to
find a better way to get the data versus scraping the data from
allareacodes.com's html pages.

Run program, an icon will show up in the System Tray. Double click it
and it will open. Enter Area Code and hit enter or click the Search
button. The Location and current time for the Area code will show up.